+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/extensions/KClassExtensions.kt
@@ -1,5 +1,7 @@
 package csense.kotlin.extensions
 
+
+//TODO create seperate package since this requires reflect
 import kotlin.reflect.*
 
 /**