+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/extensions/Nullability.kt
@@ -22,7 +22,6 @@ inline fun Any?.ifNull(crossinline action: EmptyFunction) {
  * @receiver T? the optional value
  * @param action FunctionUnit<T> the action to call if the receiver is not null
  */
-@ExperimentalContracts
 inline fun <T> T?.ifNotNull(crossinline action: FunctionUnit<T>) {
     this?.let(action)
 }