+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/types.kt
@@ -2,16 +2,17 @@
 
 package csense.kotlin
 
-/**
- * Created by kasper on 21/07/2017.
- */
+import kotlinx.coroutines.*
+
 typealias EmptyFunction = () -> Unit
 
 typealias EmptyReceiver<T> = T.() -> Unit
 typealias EmptyFunctionResult<T> = () -> T
+
 typealias AsyncEmptyFunction = suspend () -> Unit
 typealias AsyncFunctionUnit<T> = suspend (T) -> Unit
 typealias AsyncEmptyFunctionResult<T> = suspend () -> T
+
 
 /**
  * Function with 1 parameter that returns unit
@@ -21,6 +22,7 @@ typealias FunctionUnit<E> = (E) -> Unit
 typealias SuspendFunctionUnit<E> = suspend (E) -> Unit
 
 
+//region Function arity
 /**
 For library functions
  */
@@ -57,3 +59,19 @@ typealias Function5<I1, I2, I3, I4, I5, O> = (I1, I2, I3, I4, I5) -> O
  * Function with 6 inputs and potential output
  */
 typealias Function6<I1, I2, I3, I4, I5, I6, O> = (I1, I2, I3, I4, I5, I6) -> O
+//endregion
+
+//region suspend function arity
+typealias AsyncFunction0<O> = suspend () -> O
+
+typealias AsyncFunction1<I, O> = suspend (I) -> O
+typealias AsyncFunction2<I1, I2, O> = suspend (I1, I2) -> O
+typealias AsyncFunction3<I1, I2, I3, O> = suspend (I1, I2, I3) -> O
+typealias AsyncFunction4<I1, I2, I3, I4, O> = suspend (I1, I2, I3, I4) -> O
+typealias AsyncFunction5<I1, I2, I3, I4, I5, O> = suspend (I1, I2, I3, I4, I5) -> O
+typealias AsyncFunction6<I1, I2, I3, I4, I5, I6, O> = suspend (I1, I2, I3, I4, I5, I6) -> O
+//endregion
+
+
+typealias CoroutineScopeFunction = suspend CoroutineScope.() -> Unit
+typealias CoroutineScopeFunction0<O> = suspend CoroutineScope.() -> O