+++ csense-kotlin-common/src/test/kotlin/csense/kotlin/algorithms/ComparisonKtTest.kt
@@ -16,5 +16,3 @@ class ComparisonKtTest {
         Int.MIN_VALUE.toComparing().assert(Comparing.LessThan)
     }
 }
-
-