+++ csense-kotlin-test-common/src/main/kotlin/csense/kotlin/test/Async.kt
@@ -4,6 +4,7 @@ import kotlinx.coroutines.*
 
 
 fun testAsync(method: suspend () -> Unit) {
+    @Suppress("EXPERIMENTAL_API_USAGE")
     GlobalScope.launch(Dispatchers.Unconfined) {
         method()
     }