+++ csense-kotlin-test-common/src/main/kotlin/csense/kotlin/test/assertions/Comparable.kt
@@ -9,12 +9,21 @@ fun <U : Comparable<U>> U.assertLargerOrEqualTo(i: U, optMessage: String = "") {
     assertTrue(this >= i, "$this should be larger or equal to $i, but it is not.\n$optMessage")
 }
 
+fun <U : Comparable<U>> U.assertLessOrEqualTo(i: U, optMessage: String = "") {
+    assertTrue(this <= i, "$this should be smaller or equal to $i, but it is not.\n$optMessage")
+}
+
+
+fun <U : Comparable<U>> U.assertLessThan(i: U, optMessage: String = "") {
+    assertTrue(this < i, "$this should be smaller than $i, but it is not.\n$optMessage")
+}
+
 fun <U : Comparable<U>> U.assertLargerThan(i: U, optMessage: String = "") {
     assertTrue(this > i, "$this should be larger than $i, but it is not.\n$optMessage")
 }
 
 
-fun <T : Comparable<T>> T.assertNotEquals(other: T, message: String = "") {
+fun <T : Comparable<T>> T.assertNotEquals(other: T, message: String = "expected not to be equal, but $this === $other") {
     assertNotEquals(other, this, message)
 }